# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/telemetry":{"get":{"tags":["telemetry-controller"],"operationId":"getAllTelemetry","parameters":[{"name":"hostname","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TelemetryDto"}}}}}}},"post":{"tags":["telemetry-controller"],"operationId":"postTelemetry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TelemetryDto"}}}}}}},"/api/device":{"post":{"tags":["device-controller"],"operationId":"postDevice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}}}}},"/ /api/save-consumed-devices":{"post":{"tags":["extra-controller"],"operationId":"postFiveRandomDevices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDto"}}}}}}}},"/api/device/type":{"get":{"tags":["device-controller"],"operationId":"getDeviceByType","parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string","enum":["LAPTOP","DESKTOP","TABLET","SMARTPHONE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeviceDto"}}}}}}}},"/api/device/cpuUsage":{"get":{"tags":["device-controller"],"operationId":"getDeviceByCpuUsageBetween","parameters":[{"name":"upThreshold","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"lowThreshold","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DeviceDto"}}}}}}}},"components":{"schemas":{"TelemetryDto":{"type":"object","properties":{"hostname":{"type":"string"},"ip":{"type":"string"},"dataDate":{"type":"string","format":"date-time"},"hostDiskFree":{"type":"number","format":"double"},"cpuUsage":{"type":"number","format":"double"},"microphoneState":{"type":"string"},"screenCaptureAllowed":{"type":"boolean"},"audioCaptureAllowed":{"type":"boolean"}}},"DeviceDto":{"type":"object","properties":{"type":{"type":"string","enum":["LAPTOP","DESKTOP","TABLET","SMARTPHONE"]},"os":{"type":"string"},"ip":{"type":"string"},"HOSTNAME":{"type":"string"},"mac_address":{"type":"string"}}}}}}